This should now store most of it in the persistance layer
