+++ src/main/java/ch/uzh/ifi/seal/soprafs20/User/Player.java
@@ -2,13 +2,29 @@ package ch.uzh.ifi.seal.soprafs20.User;
 
 import ch.uzh.ifi.seal.soprafs20.cards.Card;
 
+import javax.persistence.*;
 import java.util.ArrayList;
+import java.util.List;
 
+@Entity
+@Table(name = "PLAYER")
 public class Player {
 
+    @Id
+    @GeneratedValue
+    private long id;
+
+    @OneToOne(targetEntity = User.class)
     User user;
+
+    @Column
     Colour colour;
-    ArrayList<Card> hand;
-    ArrayList<Figure> figures;
+
+    @OneToMany(targetEntity = Card.class)
+    List<Card> hand;
+
+    @OneToMany(targetEntity = Figure.class)
+    List<Figure> figures;
+
 
 }
