+++ src/main/java/ch/uzh/ifi/seal/soprafs20/User/User.java
@@ -1,20 +1,89 @@
 package ch.uzh.ifi.seal.soprafs20.User;
 
+import ch.uzh.ifi.seal.soprafs20.constant.UserStatus;
 
+import javax.persistence.*;
+import java.io.Serializable;
 
-import ch.uzh.ifi.seal.soprafs20.game.Game;
+import org.springframework.security.crypto.bcrypt.BCrypt;
 
-import java.util.ArrayList;
+/**
+ * Internal User Representation
+ * This class composes the internal representation of the user and defines how the user is stored in the database.
+ * Every variable will be mapped into a database field with the @Column annotation
+ * - nullable = false -> this cannot be left empty
+ * - unique = true -> this value must be unqiue across the database -> composes the primary key
+ */
+@Entity
+@Table(name = "USER")
+public class User implements Serializable {
 
-public class User {
-    Long id;
-    String username;
-    String password;
-    String token;
-    UserStatus status;
-    ArrayList<Game> games;
+    private static final long serialVersionUID = 1L;
 
-    public void createUser(){}
-    public void login(){}
-    public void logout(){}
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    @Column(nullable = false, unique = true)
+    private String email;
+
+    @Column(nullable = false, unique = true)
+    private String username;
+
+    @Column(nullable = false)
+    private String password;
+
+    @Column(nullable = false, unique = true)
+    private String token;
+
+    @Column(nullable = false)
+    private UserStatus status;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String name) {
+        this.email = name;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getToken() {
+        return token;
+    }
+
+    public void setToken(String token) {
+        this.token = token;
+    }
+
+    public UserStatus getStatus() {
+        return status;
+    }
+
+    public void setStatus(UserStatus status) {
+        this.status = status;
+    }
+
+    public boolean checkPassword(String password) {
+        return BCrypt.checkpw(password, this.password);
+    }
+
+    public void setPassword(String password) {
+        this.password = BCrypt.hashpw(password, BCrypt.gensalt());
+    }
 }
