+++ src/main/java/ch/uzh/ifi/seal/soprafs20/cards/Deck.java
@@ -1,12 +1,23 @@
 package ch.uzh.ifi.seal.soprafs20.cards;
 
+import javax.persistence.*;
+import java.io.Serializable;
 import java.util.ArrayList;
+import java.util.List;
 
-public class Deck {
-    ArrayList<Card> cards = new ArrayList<Card>();
+@Entity
+@Table(name = "DECK")
+public class Deck implements Serializable {
 
-    public Card deal(int amount){
-        //returns a Card
+    @Id
+    @GeneratedValue
+    private long id;
+
+    @OneToMany(targetEntity = Card.class)
+    List<Card> cards = new ArrayList<Card>();
+
+    public List<Card> deal(int amount){
+        //returns cards
         return null;
     }
