+++ src/main/java/ch/uzh/ifi/seal/soprafs20/cards/JokerCard.java
@@ -1,6 +1,6 @@
 package ch.uzh.ifi.seal.soprafs20.cards;
 
-public class JokerCard implements Card{
+public class JokerCard extends Card {
 
     public void transformInto(Card card){
         //maybe this should return something
