+++ src/main/java/ch/uzh/ifi/seal/soprafs20/cards/NormalCard.java
@@ -1,6 +1,6 @@
 package ch.uzh.ifi.seal.soprafs20.cards;
 
-public class NormalCard implements Card {
+public class NormalCard extends Card {
     private Suit suit;
     private Value value;
